requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.2.23
djangorestframework==3.15.1
folium==0.18.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
idna==3.10
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xyzservices==2024.9.0
